requirements.txt
numpy==1.24.1
scipy==1.9.1
open3d==0.16.1
matplotlib==3.6.2
pycairo==1.23.0
